Refactor AES Encryption
This commit fixes and refactors AES encryption.

The executables have been separated out and renamed for ease of readability and for future key types.
