+++ src/encrypt/CMakeLists.txt
@@ -6,5 +6,10 @@ find_library(cloudhsmpkcs11 STATIC)
 include_directories("../common")
 include_directories("../pkcs11_headers")
 
-add_executable(encrypt encrypt.c ecb.c cbc.c gcm.c)
-target_link_libraries(encrypt cloudhsmpkcs11)
+add_executable(aes_cbc aes_cbc.c aes.c)
+add_executable(aes_ecb aes_ecb.c aes.c)
+add_executable(aes_gcm aes_gcm.c aes.c)
+
+target_link_libraries(aes_cbc cloudhsmpkcs11)
+target_link_libraries(aes_ecb cloudhsmpkcs11)
+target_link_libraries(aes_gcm cloudhsmpkcs11)
